Remove unnecessary library linkage
Some libraries are only needed for tcpcryptd, but not for the other
tools.  Other libraries (like librt and libz) aren't needed at all and
can be dropped entirely from the build.

This is tested on a debian amd64 Linux system (debian testing/jessie).
If it turns out that some of the libraries are actually needed on
other architectures, i'm happy to help sort out patches to link them
back in for those architectures specifically.